WIP: In the process of fixing type casting issues

+++ src/main/java/de/csbd/learnathon/command/FlowComputation.java
@@ -17,67 +17,67 @@ import net.imglib2.algorithm.neighborhood.RectangleShape;
 import net.imglib2.algorithm.neighborhood.RectangleShape.NeighborhoodsAccessible;
 import net.imglib2.algorithm.region.hypersphere.HyperSphere;
 import net.imglib2.img.Img;
-import net.imglib2.img.ImgFactory;
-import net.imglib2.img.cell.CellImgFactory;
 import net.imglib2.interpolation.neighborsearch.NearestNeighborSearchInterpolatorFactory;
 import net.imglib2.neighborsearch.NearestNeighborSearch;
 import net.imglib2.neighborsearch.NearestNeighborSearchOnKDTree;
-import net.imglib2.type.numeric.real.FloatType;
+import net.imglib2.type.numeric.real.DoubleType;
 import net.imglib2.view.Views;
 
 public class FlowComputation {
 
-	public static Img< FloatType > getTMFlow( Img< FloatType > img ) {
+	public static Img< DoubleType > getTMFlow( Img< DoubleType > img ) {
 		float sigma = 2;
-		Img< FloatType > smoothed_img = gaussian_smoothing2D( img, sigma );
+		Img< DoubleType > smoothed_img = gaussian_smoothing2D( img, sigma );
 		ArrayList< LocalMaximaQuartet > localMaxima = findLocalMax( img, 5 );
 		ArrayList< LocalMaximaQuartet > thresholdedLocalMaxima = thresholdedMaxima( localMaxima, 20 );
 		ArrayList< FlowVector > sparseFlow = templateMatching( smoothed_img, thresholdedLocalMaxima );
-		Img< FloatType > denseFlow = interpolateFlowNN( sparseFlow,smoothed_img );
+		Img< DoubleType > denseFlow = interpolateFlowNN( sparseFlow, smoothed_img );
 		return denseFlow;
 	}
 
-	private static Img< FloatType > interpolateFlowNN( ArrayList< FlowVector > sparseFlow,Img< FloatType > img ) {
+	private static Img< DoubleType > interpolateFlowNN( ArrayList< FlowVector > sparseFlow, Img< DoubleType > img ) {
 
-//		final ImgFactory< FloatType > imgFactory = new CellImgFactory<>( new FloatType(), 5 );
-//		final Img< FloatType > denseFlow  = imgFactory.create( img.dimension( 0 ), img.dimension( 1 ),img.dimension( 2 )*2-2  );
+//		final ImgFactory< DoubleType > imgFactory = new CellImgFactory<>( new DoubleType(), 5 );
+//		final Img< DoubleType > denseFlow  = imgFactory.create( img.dimension( 0 ), img.dimension( 1 ),img.dimension( 2 )*2-2  );
 		
-//		final ImgFactory< FloatType > imgFactory = new CellImgFactory<>( new FloatType(), 5 );
-//		final Img< FloatType > denseFlow  = imgFactory.create( img.dimension( 0 ), img.dimension( 1 ),0  );		
+//		final ImgFactory< DoubleType > imgFactory = new CellImgFactory<>( new DoubleType(), 5 );
+//		final Img< DoubleType > denseFlow  = imgFactory.create( img.dimension( 0 ), img.dimension( 1 ),0  );		
 		
-		ArrayList<RandomAccessibleInterval< FloatType >> slices=new ArrayList<RandomAccessibleInterval< FloatType >>();
+		ArrayList< RandomAccessibleInterval< DoubleType > > slices = new ArrayList< RandomAccessibleInterval< DoubleType > >();
 		
 		for ( long pos = 0; pos < img.dimension( 2 )-1; ++pos ) {
-			 RandomAccessibleInterval< FloatType > slice = Views.hyperSlice( img, 2, pos );
+			RandomAccessibleInterval< DoubleType > slice = Views.hyperSlice( img, 2, pos );
 			
 			 ///FinalInterval interval = new FinalInterval( new long[] { 375, 200 } );
 			
-			 RealPointSampleList< FloatType > realIntervalU = new RealPointSampleList<>( 2 );
-			 RealPointSampleList< FloatType > realIntervalV = new RealPointSampleList<>( 2 );
+			RealPointSampleList< DoubleType > realIntervalU = new RealPointSampleList<>( 2 );
+			RealPointSampleList< DoubleType > realIntervalV = new RealPointSampleList<>( 2 );
 			 for ( FlowVector f : sparseFlow ) { 
 				 if (f.getT()==pos){ 
 					 RealPoint point = new RealPoint( 2 );
 					 point.setPosition(f.getX(),0);
 					 point.setPosition(f.getY(),1);
 					 
-					 realIntervalU.add(point, new FloatType(f.getU()));
-					 realIntervalV.add(point, new FloatType(f.getV()));
+					realIntervalU.add( point, new DoubleType( f.getU() ) );
+					realIntervalV.add( point, new DoubleType( f.getV() ) );
 				 }
 			 }
 			 
 			 
 			 
-			 NearestNeighborSearch< FloatType > searchU =new NearestNeighborSearchOnKDTree<>(new KDTree<>( realIntervalU ) );
-			 NearestNeighborSearch< FloatType > searchV =new NearestNeighborSearchOnKDTree<>(new KDTree<>( realIntervalV ) );
+			NearestNeighborSearch< DoubleType > searchU = new NearestNeighborSearchOnKDTree<>( new KDTree<>( realIntervalU ) );
+			NearestNeighborSearch< DoubleType > searchV = new NearestNeighborSearchOnKDTree<>( new KDTree<>( realIntervalV ) );
 			 
-			 RealRandomAccessible< FloatType > realRandomAccessibleU =Views.interpolate( searchU, new NearestNeighborSearchInterpolatorFactory< FloatType >() );
-			 RealRandomAccessible< FloatType > realRandomAccessibleV =Views.interpolate( searchV, new NearestNeighborSearchInterpolatorFactory< FloatType >() );
+			RealRandomAccessible< DoubleType > realRandomAccessibleU =
+					Views.interpolate( searchU, new NearestNeighborSearchInterpolatorFactory< DoubleType >() );
+			RealRandomAccessible< DoubleType > realRandomAccessibleV =
+					Views.interpolate( searchV, new NearestNeighborSearchInterpolatorFactory< DoubleType >() );
 			 
-			 RandomAccessible< FloatType > randomAccessibleU = Views.raster( realRandomAccessibleU );
-			 RandomAccessible< FloatType > randomAccessibleV = Views.raster( realRandomAccessibleV );
+			RandomAccessible< DoubleType > randomAccessibleU = Views.raster( realRandomAccessibleU );
+			RandomAccessible< DoubleType > randomAccessibleV = Views.raster( realRandomAccessibleV );
 			 
-			 RandomAccessibleInterval< FloatType > sliceU = Views.interval( randomAccessibleU, slice );
-			 RandomAccessibleInterval< FloatType > sliceV = Views.interval( randomAccessibleV, slice );
+			RandomAccessibleInterval< DoubleType > sliceU = Views.interval( randomAccessibleU, slice );
+			RandomAccessibleInterval< DoubleType > sliceV = Views.interval( randomAccessibleV, slice );
 			 
 			 slices.add(sliceU);
 			 slices.add(sliceV);
@@ -85,26 +85,26 @@ public class FlowComputation {
 			
 		}
 		
-		Img<FloatType> stack =(Img<FloatType>) Views.stack(slices);
+		Img< DoubleType > stack = ( Img< DoubleType > ) Views.stack( slices );
 	
 		
 		return stack;
 	}
 
-	private static ArrayList< LocalMaximaQuartet > findLocalMax( Img< FloatType > img, int r ) {
+	private static ArrayList< LocalMaximaQuartet > findLocalMax( Img< DoubleType > img, int r ) {
 
 		ArrayList< LocalMaximaQuartet > localMaxList = new ArrayList<>();
 		for ( long pos = 0; pos < img.dimension( 2 ); ++pos ) {
-			RandomAccessibleInterval< FloatType > slice = Views.hyperSlice( img, 2, pos );
-			Cursor< FloatType > cursor = Views.iterable( slice ).cursor();
+			RandomAccessibleInterval< DoubleType > slice = Views.hyperSlice( img, 2, pos );
+			Cursor< DoubleType > cursor = Views.iterable( slice ).cursor();
 
 			while ( cursor.hasNext() ) {
 				cursor.fwd();
-				HyperSphere< FloatType > smallSphere = new HyperSphere<>( slice, cursor, r );
-				FloatType centerValue = cursor.get();
+				HyperSphere< DoubleType > smallSphere = new HyperSphere<>( slice, cursor, r );
+				DoubleType centerValue = cursor.get();
 				boolean isMaximum = true;
 
-				for ( final FloatType value : smallSphere ) {
+				for ( final DoubleType value : smallSphere ) {
 					if ( centerValue.compareTo( value ) < 0 ) {
 						isMaximum = false;
 						break;
@@ -121,7 +121,7 @@ public class FlowComputation {
 		return localMaxList;
 	}
 
-	private static Img< FloatType > gaussian_smoothing2D( Img< FloatType > img, float sigma ) {
+	private static Img< DoubleType > gaussian_smoothing2D( Img< DoubleType > img, float sigma ) {
 
 		float[] s = new float[ img.numDimensions() - 1 ];
 
@@ -129,7 +129,7 @@ public class FlowComputation {
 			s[ d ] = sigma;
 
 		for ( long pos = 0; pos < img.dimension( 2 ); ++pos ) {
-			RandomAccessibleInterval< FloatType > slice = Views.hyperSlice( img, 2, pos );
+			RandomAccessibleInterval< DoubleType > slice = Views.hyperSlice( img, 2, pos );
 			Gauss3.gauss( sigma, Views.extendMirrorSingle( slice ), slice );
 		}
 
@@ -148,34 +148,34 @@ public class FlowComputation {
 		return a;
 	}
 
-	private static ArrayList< FlowVector > templateMatching( Img< FloatType > img, ArrayList< LocalMaximaQuartet > quartet ) {
+	private static ArrayList< FlowVector > templateMatching( Img< DoubleType > img, ArrayList< LocalMaximaQuartet > quartet ) {
 		RectangleShape rectangle = new RectangleShape( 5, false );
 		ArrayList<FlowVector> flowVectorList = new ArrayList<>(); 
 		int ii = 0;
 		int jj = 0;
 
 		for ( long pos = 0; pos < img.dimension( 2 )-1; ++pos ) {
-			RandomAccessibleInterval< FloatType > slice_small = Views.hyperSlice( img, 2, pos );
-			NeighborhoodsAccessible< FloatType > neighborhoods1 = rectangle.neighborhoodsRandomAccessible( slice_small );
-			RandomAccess< Neighborhood< FloatType > > ran1 = neighborhoods1.randomAccess(); 
-			RandomAccessibleInterval< FloatType > slice_big = Views.hyperSlice( img, 2, pos+1 );
-			NeighborhoodsAccessible< FloatType > neighborhoods2 = rectangle.neighborhoodsRandomAccessible( slice_big );
-			RandomAccess< Neighborhood< FloatType > > ran2 = neighborhoods2.randomAccess(); 
+			RandomAccessibleInterval< DoubleType > slice_small = Views.hyperSlice( img, 2, pos );
+			NeighborhoodsAccessible< DoubleType > neighborhoods1 = rectangle.neighborhoodsRandomAccessible( slice_small );
+			RandomAccess< Neighborhood< DoubleType > > ran1 = neighborhoods1.randomAccess();
+			RandomAccessibleInterval< DoubleType > slice_big = Views.hyperSlice( img, 2, pos + 1 );
+			NeighborhoodsAccessible< DoubleType > neighborhoods2 = rectangle.neighborhoodsRandomAccessible( slice_big );
+			RandomAccess< Neighborhood< DoubleType > > ran2 = neighborhoods2.randomAccess();
 			for ( LocalMaximaQuartet q : quartet ) {
 				if(q.getT() == pos) {
 					
 				
     				ran1.setPosition( new int [] {q.getX(), q.getY()} );
-    				Neighborhood< FloatType > n1 = ran1.get();
+					Neighborhood< DoubleType > n1 = ran1.get();
 
     				double maxVal = -Double.MAX_VALUE;
     				for (int i = -5; i< 5; i++ ) {
     					for (int j = -5; j < 5; j++) {
     						ran2.setPosition( new int [] {q.getX()+i, q.getY()+j} );
-    						Neighborhood< FloatType > n2 = ran2.get();
+							Neighborhood< DoubleType > n2 = ran2.get();
     						
-    						Cursor<FloatType> cursorA = n1.cursor();
-    				        Cursor<FloatType> cursorB = n2.cursor();
+							Cursor< DoubleType > cursorA = n1.cursor();
+							Cursor< DoubleType > cursorB = n2.cursor();
     				        double sum = 0;
     				        while (cursorA.hasNext()) {
     				            double valueA = cursorA.next().getRealDouble();
@@ -206,12 +206,12 @@ public class FlowComputation {
 			return flowVectorList;
 	}
 
-//	public static Img< FloatType > getConstantFlow( RandomAccessibleInterval< DoubleType > image ) {
+//	public static Img< DoubleType > getConstantFlow( RandomAccessibleInterval< DoubleType > image ) {
 //
-//		final ImgFactory< FloatType > imgFactory = new CellImgFactory<>( new FloatType(), 5 );
-//		final Img< FloatType > img1 = imgFactory.create( image.dimension( 0 ), image.dimension( 1 ), 8 * 2 - 2 );
+//		final ImgFactory< DoubleType > imgFactory = new CellImgFactory<>( new DoubleType(), 5 );
+//		final Img< DoubleType > img1 = imgFactory.create( image.dimension( 0 ), image.dimension( 1 ), 8 * 2 - 2 );
 //
-//		Cursor< FloatType > cursorInput = img1.cursor();
+//		Cursor< DoubleType > cursorInput = img1.cursor();
 //
 //		while ( cursorInput.hasNext() ) {
 //			cursorInput.fwd();
